Android: moves content/ to use long for JNI.
BUG=317523

Review URL: https://codereview.chromium.org/59673007

git-svn-id: svn://svn.chromium.org/chrome/trunk/src@235815 0039d316-1c4b-4281-b951-d872f2087c98

+++ content/browser/android/child_process_launcher_android.cc
@@ -71,7 +71,7 @@ static void SetSurfacePeer(
 // the ChildProcess could not be created.
 static void OnChildProcessStarted(JNIEnv*,
                                   jclass,
-                                  jint client_context,
+                                  jlong client_context,
                                   jint handle) {
   StartChildProcessCallback* callback =
       reinterpret_cast<StartChildProcessCallback*>(client_context);
@@ -123,7 +123,7 @@ void StartChildProcess(
       j_file_ids.obj(),
       j_file_fds.obj(),
       j_file_auto_close.obj(),
-      reinterpret_cast<jint>(new StartChildProcessCallback(callback)));
+      reinterpret_cast<intptr_t>(new StartChildProcessCallback(callback)));
 }
 
 void StopChildProcess(base::ProcessHandle handle) {
