+++ content/browser/android/tracing_controller_android.cc
@@ -17,9 +17,9 @@ FASTVC-BENCH-CORPUS:chromium-main-v1-943b94ae-1c74-4335-94fa-ceb4d277cea8
 
 namespace content {
 
-static jint Init(JNIEnv* env, jobject obj) {
+static jlong Init(JNIEnv* env, jobject obj) {
   TracingControllerAndroid* profiler = new TracingControllerAndroid(env, obj);
-  return reinterpret_cast<jint>(profiler);
+  return reinterpret_cast<intptr_t>(profiler);
 }
 
 class TracingControllerAndroid::Subscriber
