+++ content/public/android/java/src/org/chromium/content/browser/ContentView.java
@@ -51,7 +51,7 @@ public class ContentView extends FrameLayout
      * @param windowAndroid An instance of the WindowAndroid.
      * @return A ContentView instance.
      */
-    public static ContentView newInstance(Context context, int nativeWebContents,
+    public static ContentView newInstance(Context context, long nativeWebContents,
             WindowAndroid windowAndroid) {
         return newInstance(context, nativeWebContents, windowAndroid, null,
                 android.R.attr.webViewStyle);
@@ -66,7 +66,7 @@ public class ContentView extends FrameLayout
      * @param attrs The attributes of the XML tag that is inflating the view.
      * @return A ContentView instance.
      */
-    public static ContentView newInstance(Context context, int nativeWebContents,
+    public static ContentView newInstance(Context context, long nativeWebContents,
             WindowAndroid windowAndroid, AttributeSet attrs) {
         // TODO(klobag): use the WebViewStyle as the default style for now. It enables scrollbar.
         // When ContentView is moved to framework, we can define its own style in the res.
@@ -84,7 +84,7 @@ public class ContentView extends FrameLayout
      * @param defStyle The default style to apply to this view.
      * @return A ContentView instance.
      */
-    public static ContentView newInstance(Context context, int nativeWebContents,
+    public static ContentView newInstance(Context context, long nativeWebContents,
             WindowAndroid windowAndroid, AttributeSet attrs, int defStyle) {
         if (Build.VERSION.SDK_INT < Build.VERSION_CODES.JELLY_BEAN) {
             return new ContentView(context, nativeWebContents, windowAndroid, attrs, defStyle);
@@ -94,7 +94,7 @@ public class ContentView extends FrameLayout
         }
     }
 
-    protected ContentView(Context context, int nativeWebContents, WindowAndroid windowAndroid,
+    protected ContentView(Context context, long nativeWebContents, WindowAndroid windowAndroid,
             AttributeSet attrs, int defStyle) {
         super(context, attrs, defStyle);
