+++ content/public/android/java/src/org/chromium/content/browser/GenericTouchGesture.java
@@ -31,7 +31,7 @@ public class GenericTouchGesture {
 
     private TimeAnimator mTimeAnimator;
 
-    private int mNativePtr;
+    private long mNativePtr;
     private long mDownTime;
     private long mStartTime;
 
@@ -173,7 +173,7 @@ public class GenericTouchGesture {
     }
 
     @CalledByNative
-    void start(int nativePtr) {
+    void start(long nativePtr) {
         assert mNativePtr == 0;
         mNativePtr = nativePtr;
         mStartTime = SystemClock.uptimeMillis();
@@ -292,6 +292,6 @@ public class GenericTouchGesture {
     }
 
     private native float nativeGetDelta(
-            int nativeGenericTouchGestureAndroid, float scale);
-    private native void nativeSetHasFinished(int nativeGenericTouchGestureAndroid);
+            long nativeGenericTouchGestureAndroid, float scale);
+    private native void nativeSetHasFinished(long nativeGenericTouchGestureAndroid);
 }
