+++ content/public/android/java/src/org/chromium/content/browser/InterstitialPageDelegateAndroid.java
@@ -14,7 +14,7 @@ import org.chromium.base.JNINamespace;
 @JNINamespace("content")
 public class InterstitialPageDelegateAndroid {
 
-    private int mNativePtr;
+    private long mNativePtr;
 
     /**
      * Constructs an interstitial with the given HTML content.
@@ -28,7 +28,7 @@ public class InterstitialPageDelegateAndroid {
     /**
      * @return The pointer to the underlying native counterpart.
      */
-    public int getNative() {
+    public long getNative() {
         return mNativePtr;
     }
 
@@ -74,7 +74,7 @@ public class InterstitialPageDelegateAndroid {
         if (mNativePtr != 0) nativeDontProceed(mNativePtr);
     }
 
-    private native int nativeInit(String htmlContent);
-    private native void nativeProceed(int nativeInterstitialPageDelegateAndroid);
-    private native void nativeDontProceed(int nativeInterstitialPageDelegateAndroid);
+    private native long nativeInit(String htmlContent);
+    private native void nativeProceed(long nativeInterstitialPageDelegateAndroid);
+    private native void nativeDontProceed(long nativeInterstitialPageDelegateAndroid);
 }
