+++ content/public/android/java/src/org/chromium/content/browser/accessibility/JellyBeanBrowserAccessibilityManager.java
@@ -23,7 +23,7 @@ import java.util.List;
 public class JellyBeanBrowserAccessibilityManager extends BrowserAccessibilityManager {
     private AccessibilityNodeProvider mAccessibilityNodeProvider;
 
-    JellyBeanBrowserAccessibilityManager(int nativeBrowserAccessibilityManagerAndroid,
+    JellyBeanBrowserAccessibilityManager(long nativeBrowserAccessibilityManagerAndroid,
             ContentViewCore contentViewCore) {
         super(nativeBrowserAccessibilityManagerAndroid, contentViewCore);
